Change wording when unable to join another server
*Note I may be misunderstanding the complete use of this message, anyway,

For example: When the /server command is run on an offline server, That message is shown. Having the word "default" (to me) is confusing, as is it trying to tell me it couldn't connect and it couldn't move me to a default server? That's why i made this clarification change.