Add ping throttle
Sending a handshake and a bunch of StatusRequests is a very effective
DDoS attack because each SR is just one x00 byte and the StatusRequest
handler is quite CPU-intensive. This PR adds a configurable limit to
disconnect a client after they have pinged a certain amount of times per
connection. Setting it to -1 disables this throttle. Defaults to 5
because I can't think of any use case beyond two and it appears to
reasonably protect against this exploit.