Add a mechanism for rejecting old Forge clients.
We check the mod list that the client sends for the FML version it is running, specifically, the build number, via a Regular Expression. If the FML build number is 0 (which usually indicates a testing build) or above the min build required (currently set to 1209, but will be updated when the FML reset patch is pulled), the connection proceeds as normal. Otherwise any connection to a Forge server will be rejected, as switching servers would otherwise crash the client.

This will disconnect users connecting to the default server, and sending a message to users who try to switch in game. We are looking to see if there is a good way to make users connecting to the first/default server connect to the fallback server instead.