Fixed Forge's DimensionRegisterPacket being ignored after handshake.
After a client switched to a plugin-created dimension and restarted client,
the next login would result in a crash. This was caused by Forge's
DimensionRegisterPacket being ignored during the "DONE" phase of our
ForgeServerHandshakeState class. To fix this issue and others like it,
we now pass all packets directly to client after handshake is done.

More information about this issue can be found here

https://github.com/MinecraftPortCentral/Cauldron-Issues/issues/4

* Added Forge login support for dimension ID's > 127.