Add support for FML with IP Forwarding enabled
FML adds a \00FML\00 marker to the host field, so Forge can determine whether or not to start a Forge handshake, making way to allow vanilla clients to connect to Forge servers that don't need a client modification. However, Bungee also uses the field, and the two implementations collide when using Spigot.

The original fix was to not send the FML information at the same time as the IP forwarding, you could have one or the other, but not both. This was OK, as no FML servers supported IP forwarding as of time of the patch. This was implemented in commit 4809f1f80ace9ae87b91453c8887c70f5e098bd0.

However, there is now at least one Forge coremod that intends to support IP forwarding. To be able to support Forge with IP forwarding, a way to be able to support the FML token (and any other host data) is needed. This patch adds the extra data onto the end of the IP Forwarding host field, separating the IP forwarding data from the other data using the character combination of "\00|" - the pipe being chosen as it will not appear at the beginning of any IP forwarding field and so will leave the IP forwarding fields intact.

Implementations MUST be aware of this as this is potentially a breaking change - and should split on this marker (but only the first time it's seen) before doing IP forwarding processing.

The host packet thus becomes "<host>\0<IP>\0<UUID>\0<JSON encoded GameProfile>\0|<Extra Data>".