Comply with international naming convention on colour/color
From http://stackoverflow.com/a/2444682/2479087 but copied here in case it is edited. This argument is for C++ but apply to Java as well:
"You should use the American spelling of color. 99% of code out there uses this, even including much of the code written by British or Australian English speakers. If you use something different, you (or someone else that uses your code) are just going to end up forgetting which to use at some point and making a needless mistake."