Add method to translate hex color codes
Added a method to the ChatColor class which allows you to choose a start and end tag, for example, start: "#<", end: ">" and take any hex code inside the tags and translate it to a color code. It then returns the new string with anything from the beginning of the start tag to the end of the end tag replaced with the hex color code. This should be added because, at the moment, the hex color code requires developers to create either a class or a method just to use custom messages with hex colors. This would make a lot more sense if it was included the official API.