Enhance permission checking during server connects
Previously the only time server permissions were tested was using the /server command.
This left a door for every ProxiedPlayer#connect(ServerInfo) invocation to ignore the ServerInfo#canAccess(ProxiedPlayer) check. This PR implements a few features:
- a config option to always go through the canAccess check during any server connect attempt;
  - I disabled this by default as to retain current behaviour and to not disrupt anyones current setups.
- An option to override whatever setting had been set in the config by directly altering this on the ServerConnectRequest.
- A new ServerConnectRequest REASON field 'UNAUTHORISED'

I feel like this is a feature needed for security purposes as to prevent accidental entrances to hidden servers. If needed of course plugins can bypass the flags.