Use Tree{Map,Set} for config options
Uses TreeMap and TreeSet for various configuration options such that the
options are ordered consistently. The previous HashMap implementation
would cause the config options to get shuffled around in the config
files every time bungeecord was shut down.

Still uses the HashMap-based implementation for non-config uses, as it
is more performant.

Closes #2013.