Cache MessageFormates for Translation
The default implementation of java ( MessageFormat.format ) create a new MessageFormat every time its executed. This behavior is wasting much performance as it creates a new pettern every time with the applyPattern(String) method of the MessageFormat class. To increase the performace of the getTranslation method and make the process way faster I added a simple cache for the MessageFormats so it is not creating a new one every time we want to transalte