Fix intent issues in InitialHandler
It is possible to make the server call an infinit amount of events, while an event-callback waits for all Intents to be completed, to change to a new state.

Adding an State for Ping and for Login Events is the simplest way to fix this.

This fixes multiple sending of legacy packets, it is possible to send LegacyPing and while the event call is waiting for all intents to be complete send an Legacy Handshake or a normal handshake and establish a real connection while the legacy ping event is running in the background, because the state is handshake and the channel not closed instanly.

It is also possible to call the PreLoginEvent infinite times, while the first one is waiting for a registered intend, wich could potentially crash the server, if abused.

Also it was possible to send a Ping Packet to the Server befor the server sent the StatusResponse to the client if an intend is registered for the ServerPingEvent. This should not be possible.