Fix slow byte array access in native code
"GetByteArrayElements" will usually copy the array out of the JVM's heap to a temporary buffer first for safety before allowing it to be accessed, which is pointless if all you're doing is preparing arguments for a memcpy. "GetByteArrayRegion" can memcpy the data in the array directly from the JVM heap to a destination address without allocating a useless intermediate copy first.