Add netty signal when compression state is changed
Third party plugins using the API provided in https://github.com/SpigotMC/BungeeCord/pull/3776 need to reorder their own pipeline elements as soon as the compression is enabled, which is currently not possible (at least not in a proper way without delaying packets). This PR adds a simple netty signal which can be used for this very use-case and is provided similar in other proxy/server projects as well.