#3885: Remove manual player removal from tab list
It seems like my changes in #3656 made this redundant.

I could not reproduce what was described in https://github.com/SpigotMC/BungeeCord/commit/830f18a

I guess the reason for that is that in my change i used the offline uuid for rewriting if ip forward is disable and online mode enabled.

In exact this case before the premium uuid was used and could later not be rewritten anymore when the player has disconnected.